% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webshot2-package.R
\docType{package}
\name{webshot2-package}
\alias{webshot2}
\alias{webshot2-package}
\title{webshot2: Take Screenshots of Web Pages}
\description{
Takes screenshots of web pages, including Shiny applications and R Markdown documents. 'webshot2' uses headless Chrome or Chromium as the browser back-end.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/webshot2/}
  \item \url{https://github.com/rstudio/webshot2}
  \item Report bugs at \url{https://github.com/rstudio/webshot2/issues}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@posit.co}

Other contributors:
\itemize{
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID}) [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
